+++ app/src/main/java/com/cmput301f17t11/cupofjava/Models/BottomNavigationViewHelper.java
@@ -7,7 +7,7 @@
  * Copyright (c) 2017 Cup Of Java. All rights reserved.
  */
 
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Models;
