+++ app/src/main/java/com/cmput301f17t11/cupofjava/Models/Habit.java
@@ -7,11 +7,10 @@
  * Copyright (c) 2017 Cup Of Java. All rights reserved.
  */
 
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Models;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.Calendar;
 
 import io.searchbox.annotations.JestId;
