+++ app/src/main/java/com/cmput301f17t11/cupofjava/Models/HabitEvent.java
@@ -7,8 +7,10 @@
  * Copyright (c) 2017 Cup Of Java. All rights reserved.
  */
 
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Models;
 
+
+import com.cmput301f17t11.cupofjava.Models.Habit;
 
 import java.io.Serializable;
 import java.text.SimpleDateFormat;
