+++ app/src/main/java/com/cmput301f17t11/cupofjava/Models/HabitEventHistory.java
@@ -7,7 +7,10 @@
  * Copyright (c) 2017 Cup Of Java. All rights reserved.
  */
 
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Models;
+
+import com.cmput301f17t11.cupofjava.Models.Habit;
+import com.cmput301f17t11.cupofjava.Models.HabitEvent;
 
 import java.io.Serializable;
 import java.util.ArrayList;
