+++ app/src/main/java/com/cmput301f17t11/cupofjava/Models/HomePagerAdapter.java
@@ -1,8 +1,11 @@
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Models;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
+
+import com.cmput301f17t11.cupofjava.Views.NearbyTab;
+import com.cmput301f17t11.cupofjava.Views.HabitEventTimeLineActivity;
 
 /**
  * Created by Moe on 2017-11-28.
