+++ app/src/main/java/com/cmput301f17t11/cupofjava/Models/UserList.java
@@ -1,10 +1,9 @@
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Models;
+
+import com.cmput301f17t11.cupofjava.Models.User;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.IllegalFormatCodePointException;
 
 
 /**
