+++ app/src/main/java/com/cmput301f17t11/cupofjava/Views/FollowingTab.java
@@ -1,7 +1,5 @@
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Views;
 
-import android.content.Context;
-import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -9,16 +7,18 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.cmput301f17t11.cupofjava.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link FollowersTab.OnFragmentInteractionListener} interface
+ * {@link FollowingTab.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link FollowersTab#newInstance} factory method to
+ * Use the {@link FollowingTab#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class FollowersTab extends Fragment {
+public class FollowingTab extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -30,7 +30,7 @@ public class FollowersTab extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public FollowersTab() {
+    public FollowingTab() {
         // Required empty public constructor
     }
 
@@ -40,11 +40,11 @@ public class FollowersTab extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment FollowersTab.
+     * @return A new instance of fragment FollowingTab.
      */
     // TODO: Rename and change types and number of parameters
-    public static FollowersTab newInstance(String param1, String param2) {
-        FollowersTab fragment = new FollowersTab();
+    public static FollowingTab newInstance(String param1, String param2) {
+        FollowingTab fragment = new FollowingTab();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -65,8 +65,7 @@ public class FollowersTab extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_followers_tab, container, false);
-
+        return inflater.inflate(R.layout.fragment_following_tab, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
@@ -108,6 +107,4 @@ public class FollowersTab extends Fragment {
         // TODO: Update argument type and name
         void onFragmentInteraction(Uri uri);
     }
-
-
 }
