+++ app/src/main/java/com/cmput301f17t11/cupofjava/Views/RequestsTab.java
@@ -1,6 +1,5 @@
-package com.cmput301f17t11.cupofjava;
+package com.cmput301f17t11.cupofjava.Views;
 
-import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -8,16 +7,18 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.cmput301f17t11.cupofjava.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link FollowingTab.OnFragmentInteractionListener} interface
+ * {@link RequestsTab.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link FollowingTab#newInstance} factory method to
+ * Use the {@link RequestsTab#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class FollowingTab extends Fragment {
+public class RequestsTab extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -29,7 +30,7 @@ public class FollowingTab extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public FollowingTab() {
+    public RequestsTab() {
         // Required empty public constructor
     }
 
@@ -39,11 +40,11 @@ public class FollowingTab extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment FollowingTab.
+     * @return A new instance of fragment RequestsTab.
      */
     // TODO: Rename and change types and number of parameters
-    public static FollowingTab newInstance(String param1, String param2) {
-        FollowingTab fragment = new FollowingTab();
+    public static RequestsTab newInstance(String param1, String param2) {
+        RequestsTab fragment = new RequestsTab();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -64,7 +65,7 @@ public class FollowingTab extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_following_tab, container, false);
+        return inflater.inflate(R.layout.fragment_requests_tab, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
