+++ app/src/main/res/layout/activity_add_new_habit.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t11.cupofjava.NewHabitActivity">
+    tools:context="com.cmput301f17t11.cupofjava.Views.NewHabitActivity">
 
 
     <LinearLayout
