+++ app/src/main/res/layout/activity_all_habit_view.xml
@@ -11,7 +11,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.cmput301f17t11.cupofjava.AllHabitViewActivity">
+        tools:context="com.cmput301f17t11.cupofjava.Views.AllHabitViewActivity">
 
         <LinearLayout
             android:layout_width="match_parent"
