+++ app/src/main/res/layout/activity_habit_detail_view.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t11.cupofjava.HabitDetailViewActivity">
+    tools:context="com.cmput301f17t11.cupofjava.Views.HabitDetailViewActivity">
 
 <LinearLayout
     android:layout_width="match_parent"
