+++ app/src/main/res/layout/activity_habit_time_line.xml
@@ -10,7 +10,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.cmput301f17t11.cupofjava.HabitEventTimeLineActivity">
+        tools:context="com.cmput301f17t11.cupofjava.Views.HabitEventTimeLineActivity">
 
         <LinearLayout
             android:layout_width="match_parent"
