+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:id="@+id/frame"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t11.cupofjava.MainActivity">
+    tools:context="com.cmput301f17t11.cupofjava.Views.MainActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
