+++ app/src/main/res/layout/activity_new_habit_event.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t11.cupofjava.NewHabitEventActivity">
+    tools:context="com.cmput301f17t11.cupofjava.Views.NewHabitEventActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
