+++ app/src/main/res/layout/activity_social.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.cmput301f17t11.cupofjava.SocialFragment">
+    tools:context="com.cmput301f17t11.cupofjava.Views.SocialFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
