+++ app/src/main/res/layout/activity_view_habit_event.xml
@@ -6,7 +6,7 @@
     android:layout_marginTop="4dp"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.cmput301f17t11.cupofjava.ViewHabitEventActivity">
+    tools:context="com.cmput301f17t11.cupofjava.Views.ViewHabitEventActivity">
 
     <TextView
         android:id="@+id/habitEventDetailHeadingTextView"
