+++ app/src/main/res/layout/fragment_following_tab.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.cmput301f17t11.cupofjava.FollowingTab">
+    tools:context="com.cmput301f17t11.cupofjava.Views.FollowingTab">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
