+++ app/src/main/res/layout/profile_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t11.cupofjava.ProfileTab">
+    tools:context="com.cmput301f17t11.cupofjava.Views.ProfileTab">
 
     <LinearLayout
         android:layout_width="match_parent"
